tests/submission/Microsoft/forms/placeholder.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>placeholder on input element</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta name="description" content="placeholder atrribute on elements" />
    </head>
    <body>
        <script src="/resources/testharness.js" ></script>
        <script src="/resources/testharnessreport.js" ></script>

        <div id=log></div>
    
        <div style="display : none">
        <input type="text" placeholder="placeholder text"></input>
        <input type="text" placeholder></input>
        <textarea type="text" placeholder="placeholder text"></textarea>
        </div>
        
        <script type="text/javascript">
        
        test(function() {assert_equals(document.getElementsByTagName("input")[0].placeholder, "placeholder text")}, "placeholder support on input element");
        test(function() {assert_equals(document.getElementsByTagName("input")[1].placeholder, "")}, "default value for placeholder on input element");
        test(function() {assert_equals(document.getElementsByTagName("input")[1].getAttribute("placeholder"), "")}, "default value for placeholder when retrieved using getAttribute");
        test(function() {assert_equals(document.getElementsByTagName("textarea")[0].placeholder, "placeholder text")}, "placeholder support on textarea element");
        
        </script>
    </body>
</html>
